chore: clean up contexts
