+++ src/contexts/PlayControllerContext.ts
@@ -1,8 +1,7 @@
 import React from 'react'
 import { Chess } from 'chess.ts'
-import { GameTree } from 'src/types'
+import { GameTree, BaseTreeControllerContext } from 'src/types'
 import { usePlayController } from 'src/hooks/usePlayController'
-import { BaseTreeControllerContext } from './BaseTreeControllerContext'
 
 export interface IPlayControllerContext extends BaseTreeControllerContext {
   game: ReturnType<typeof usePlayController>['game']
